+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Versions
 
-## 2.2.17
+## 2.2.19
+DateStr renamed to SentDate and stored as number; SendDate gets updated header('date') or if missing from internalDate (gmail timestmap) 
+further addes isUnread boolean to IEmail, based on the labelID = "unread"
+
+## 2.2.18
 updates klingsten-snippets dependency to 1.0.25 to bugfix @klingsten-snippets @String.splitNameFromEmail 
 
 ## 2.2.16